Move logger configurations and operations to files
This commit remove the logger (SD and configurations) from main.c and
usrconf.h to specific files logger.h and logger.c
