+++ usr/user.mk
@@ -1,8 +1,8 @@
 USERLIB = ./usr
 
-# List of all the Userlib files
-USERSRC =  $(USERLIB)/src/
-
 # Required include directories
 USERINC =  $(USERLIB) \
            $(USERLIB)/inc
+           
+# List of all the Userlib files
+USERSRC =  $(USERLIB)/src/logger.c
